requirements.txt
dj-database-url==0.4.2
Django==2.0.2
django-crispy-forms==1.7.0
django-filebrowser-no-grappelli==3.7.4
django-tinymce4-lite==1.7.0
jsmin==2.2.2
python-decouple==3.1
pytz==2018.3
